Split one large file
